docker/pyproject.toml
[project]
name = "charts-api"
version = "0.1.0"
description = "FastAPI server for deploying MCP servers and agents to Kubernetes via ArgoCD"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "kubernetes>=31.0.0",
    "apscheduler>=3.10.4",
    "httpx>=0.27.0",
    "structlog>=24.4.0",
    "pyyaml>=6.0.0",
]

[tool.ruff]
line-length = 120
